Implement slack hooks. Resolves #120. (#233)
