+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "workbench.colorCustomizations": {
+    "activityBar.background": "#202F3C",
+    "titleBar.activeBackground": "#2D4253",
+    "titleBar.activeForeground": "#F8FAFB"
+  }
+}
