+++ packages/dashboard/server/index.js
@@ -3,15 +3,15 @@ const { app } = require('./app');
 const { PORT } = require('./config');
 
 async function main() {
-  app.on('error', error => {
+  app.on('error', (error) => {
     throw error;
   });
   app.listen(PORT, () => {
-    console.log(`Listening on ${PORT}...`);
+    console.log(`Listening on http://localhost:${PORT}...`);
   });
 }
 
-main().catch(error => {
+main().catch((error) => {
   console.error(error);
   process.exit(1);
 });
