+++ packages/director/src/lib/hooks/cleanHooksData.ts
@@ -1,0 +1,12 @@
+import { hookReportSchema } from '@src/lib/schemas';
+import Ajv from 'ajv';
+import { cloneDeep } from 'lodash';
+
+const ajv = new Ajv({ removeAdditional: 'all' });
+const cleanHookReportData = ajv.compile(hookReportSchema);
+
+export const getCleanHookReportData = (data: unknown) => {
+  const cloned = cloneDeep(data);
+  cleanHookReportData(cloned);
+  return cloned;
+};
